feat: custom live server

+++ src/node_trans_session.js
@@ -10,6 +10,10 @@ const { spawn } = require('child_process');
 const dateFormat = require('dateformat');
 const mkdirp = require('mkdirp');
 const fs = require('fs');
+const { Stream, Record } = require('./models');
+const {Op} = require("sequelize");
+const os = require("os");
+const moment = require("moment-timezone");
 
 const isHlsFile = (filename) => filename.endsWith('.ts') || filename.endsWith('.m3u8')
 const isTemFiles = (filename) => filename.endsWith('.mpd') || filename.endsWith('.m4s') || filename.endsWith('.tmp')
@@ -17,6 +21,7 @@ const isTemFiles = (filename) => filename.endsWith('.mpd') || filename.endsWith(
 class NodeTransSession extends EventEmitter {
   constructor(conf) {
     super();
+    this.isRunning = false;
     this.conf = conf;
     this.getConfig = (key = null) => {
       if (!key) return
@@ -26,11 +31,30 @@ class NodeTransSession extends EventEmitter {
     }
   }
 
-  run() {
+  async run() {
+    if (this.isRunning) {
+      console.log('---------- session already running --------------');
+      return;
+    }
+    if (this.conf.mp4) {
+      this.stream = await Stream.findOne({
+        where: {
+          streamPath: this.conf.streamPath,
+          startAt: {
+            [Op.lte]: new Date(),
+          },
+          endAt: {
+            [Op.gt]: new Date(),
+          }
+        },
+        order: [['startAt', 'DESC']],
+      });
+    }
     let vc = this.conf.vc || 'copy';
     let ac = this.conf.ac || 'copy';
     let inPath = 'rtmp://127.0.0.1:' + this.conf.rtmpPort + this.conf.streamPath;
     let ouPath = `${this.conf.mediaroot}/${this.conf.streamApp}/${this.conf.streamName}`;
+    let mp4ouPath = `${this.conf.mp4root}/${this.conf.streamApp}/${this.conf.streamName}`;
     let mapStr = '';
 
     if (this.conf.rtmp && this.conf.rtmpApp) {
@@ -42,12 +66,16 @@ class NodeTransSession extends EventEmitter {
         Logger.log('[Transmuxing RTMP] ' + this.conf.streamPath + ' to ' + rtmpOutput);
       }
     }
-    if (this.conf.mp4) {
+    if (this.conf.mp4 && this.stream) {
       this.conf.mp4Flags = this.conf.mp4Flags ? this.conf.mp4Flags : '';
       let mp4FileName = dateFormat('yyyy-mm-dd-HH-MM-ss') + '.mp4';
-      let mapMp4 = `${this.conf.mp4Flags}${ouPath}/${mp4FileName}|`;
+      let mapMp4 = `${this.conf.mp4Flags}${mp4ouPath}/${mp4FileName}|`;
       mapStr += mapMp4;
-      Logger.log('[Transmuxing MP4] ' + this.conf.streamPath + ' to ' + ouPath + '/' + mp4FileName);
+      Logger.log('[Transmuxing MP4] ' + this.conf.streamPath + ' to ' + mp4ouPath + '/' + mp4FileName);
+      // create record
+      mkdirp.sync(mp4ouPath);
+      // create record
+      this.createRecord(this.stream.id, `${mp4ouPath}/${mp4FileName}`);
     }
     if (this.conf.hls) {
       this.conf.hlsFlags = this.getConfig('hlsFlags') || '';
@@ -63,6 +91,10 @@ class NodeTransSession extends EventEmitter {
       mapStr += mapDash;
       Logger.log('[Transmuxing DASH] ' + this.conf.streamPath + ' to ' + ouPath + '/' + dashFileName);
     }
+    if (mapStr === '') {
+      // output not found
+      return ;
+    }
     mkdirp.sync(ouPath);
     let argv = ['-y', '-i', inPath];
     Array.prototype.push.apply(argv, ['-c:v', vc]);
@@ -71,7 +103,10 @@ class NodeTransSession extends EventEmitter {
     Array.prototype.push.apply(argv, this.conf.acParam);
     Array.prototype.push.apply(argv, ['-f', 'tee', '-map', '0:a?', '-map', '0:v?', mapStr]);
     argv = argv.filter((n) => { return n; }); //去空
-    
+
+    //
+    this.isRunning = true;
+
     this.ffmpeg_exec = spawn(this.conf.ffmpeg, argv);
     this.ffmpeg_exec.on('error', (e) => {
       Logger.ffdebug(e);
@@ -87,19 +122,25 @@ class NodeTransSession extends EventEmitter {
 
     this.ffmpeg_exec.on('close', (code) => {
       Logger.log('[Transmuxing end] ' + this.conf.streamPath);
+      this.isRunning = false;
       this.emit('end');
       this.cleanTempFiles(ouPath)
       this.deleteHlsFiles(ouPath)
+      //async combine mp4 files
+      this.combineMp4Files();
     });
   }
 
   end() {
-    this.ffmpeg_exec.kill();
+    if (this.ffmpeg_exec) {
+      this.ffmpeg_exec.kill();
+    }
   }
 
   // delete hls files
   deleteHlsFiles (ouPath) {
     if ((!ouPath && !this.conf.hls) || this.getConfig('hlsKeep')) return
+    if (!this.conf.hls && !this.conf.dash) return
     fs.readdir(ouPath, function (err, files) {
       if (err) return
       files.filter((filename) => isHlsFile(filename)).forEach((filename) => {
@@ -111,11 +152,94 @@ class NodeTransSession extends EventEmitter {
   // delete the other files
   cleanTempFiles (ouPath) {
     if (!ouPath) return
+    if (!this.conf.hls && !this.conf.dash) return
     fs.readdir(ouPath, function (err, files) {
       if (err) return
       files.filter((filename) => isTemFiles(filename)).forEach((filename) => {
         fs.unlinkSync(`${ouPath}/${filename}`);
       });
+    });
+  }
+
+  genRandom() {
+    return  Math.random().toString(36).replace(/[^a-z]+/g, '');
+  }
+
+  createRecord(streamId, fileName) {
+    return Record.create({
+      streamId: streamId,
+      fileName: fileName, // full path
+    });
+  }
+
+  getStreamRecords(streamId) {
+      return Record.findAll({
+          where: {
+              streamId: streamId,
+          },
+          order: [['createdAt', 'ASC']]
+      }).then(results => {
+          return results.map(record => record.fileName);
+      });
+  }
+
+  async combineMp4Files() {
+    if (!this.conf.mp4 || !this.stream) {
+      return;
+    }
+    // mp4 ou path
+    let mp4ouPath = `${this.conf.mp4root}/${this.conf.streamApp}/${this.conf.streamName}`;
+    let mp4S3ouPath = this.conf.mp4S3root ? `${this.conf.mp4S3root}/${this.conf.streamApp}/${this.conf.streamName}` : null;
+
+    // ouFile
+    const fileName = moment(this.stream.startAt).tz('Asia/Ulaanbaatar').format('YYYY-MM-DD-HH-mm-00');
+    const ouTmpFile = `${mp4ouPath}/${this.genRandom()}.mp4`;
+    const ouFile = `${mp4ouPath}/cache-${fileName}.mp4`;
+    // s3 storage path
+    const ouS3File = this.conf.mp4S3root ? `${mp4S3ouPath}/cache-${fileName}.mp4`: null;
+    // sync s3 folder
+    if (this.conf.mp4S3root){
+      mkdirp.sync(`${this.conf.mp4S3root}/${this.conf.streamApp}/${this.conf.streamName}`);
+    }
+    // mp4 file list
+    let files = await this.getStreamRecords(this.stream.id);
+    files = files.filter(path => fs.existsSync(path));
+
+    const tmpFileList = `${os.tmpdir()}/${this.genRandom()}`;
+    files.forEach(file => {
+      fs.appendFileSync(tmpFileList, `file ${file}` + os.EOL);
+    });
+
+    let argv = [];
+    Array.prototype.push.apply(argv, ['-f', 'concat']);
+    Array.prototype.push.apply(argv, ['-safe', '0']);
+    Array.prototype.push.apply(argv, ['-i', tmpFileList]);
+    Array.prototype.push.apply(argv, ['-c', 'copy']);
+    Array.prototype.push.apply(argv, ['-y', ouTmpFile]);
+    //argv = argv.filter((n) => { return n });
+    // execute ffmpeg command
+    const ffmpeg_exec = spawn(this.conf.ffmpeg, argv, { shell: true });
+    ffmpeg_exec.on('error', (e) => {
+      Logger.ffdebug(e);
+    });
+    ffmpeg_exec.stdout.on('data', (data) => {
+      Logger.ffdebug(`[ConcatRecord]：${data}`);
+    });
+    ffmpeg_exec.stderr.on('data', (data) => {
+      Logger.ffdebug(`[ConcatRecord]：${data}`);
+    });
+    ffmpeg_exec.on('close', (code) => {
+      Logger.log('[ConcatRecord] Closed:' + this.conf.streamPath);
+
+      // rename to target file
+      fs.renameSync(ouTmpFile, ouFile);
+
+      // copy to s3 file system
+      if (ouS3File) {
+        fs.copyFileSync(ouFile, ouS3File);
+      }
+      // unlink file list
+      fs.unlinkSync(tmpFileList);
     });
   }
 }
